Update eslint to v9
