+++ package.json
@@ -67,9 +67,10 @@
   },
   "devDependencies": {
     "@eslint-community/eslint-plugin-eslint-comments": "4.5.0",
+    "@eslint/compat": "1.3.2",
     "@playwright/test": "1.55.0",
     "@stoplight/spectral-cli": "6.15.0",
-    "@stylistic/eslint-plugin-js": "3.1.0",
+    "@stylistic/eslint-plugin": "5.3.1",
     "@stylistic/stylelint-plugin": "4.0.0",
     "@types/codemirror": "5.60.16",
     "@types/dropzone": "5.7.9",
@@ -82,24 +83,24 @@
     "@types/throttle-debounce": "5.0.2",
     "@types/tinycolor2": "1.4.6",
     "@types/toastify-js": "1.12.4",
-    "@typescript-eslint/eslint-plugin": "8.43.0",
     "@typescript-eslint/parser": "8.43.0",
     "@vitejs/plugin-vue": "6.0.1",
     "@vitest/eslint-plugin": "1.3.9",
-    "eslint": "8.57.0",
+    "eslint": "9.35.0",
     "eslint-import-resolver-typescript": "4.4.4",
-    "eslint-plugin-array-func": "4.0.0",
-    "eslint-plugin-github": "5.0.2",
+    "eslint-plugin-array-func": "5.0.2",
+    "eslint-plugin-github": "6.0.0",
     "eslint-plugin-import-x": "4.16.1",
     "eslint-plugin-no-jquery": "3.1.1",
-    "eslint-plugin-no-use-extend-native": "0.5.0",
+    "eslint-plugin-no-use-extend-native": "0.7.2",
     "eslint-plugin-playwright": "2.2.2",
     "eslint-plugin-regexp": "2.10.0",
     "eslint-plugin-sonarjs": "3.0.5",
-    "eslint-plugin-unicorn": "56.0.1",
+    "eslint-plugin-unicorn": "61.0.2",
     "eslint-plugin-vue": "10.4.0",
     "eslint-plugin-vue-scoped-css": "2.12.0",
     "eslint-plugin-wc": "3.0.1",
+    "globals": "16.4.0",
     "happy-dom": "18.0.1",
     "markdownlint-cli": "0.45.0",
     "material-icon-theme": "5.27.0",
@@ -112,6 +113,7 @@
     "stylelint-value-no-unknown-custom-properties": "6.0.1",
     "svgo": "4.0.0",
     "type-fest": "4.41.0",
+    "typescript-eslint": "8.43.0",
     "updates": "16.7.0",
     "vite-string-plugin": "1.4.6",
     "vitest": "3.2.4",
