+++ types.d.ts
@@ -1,4 +1,33 @@
 declare module '@techknowlogick/license-checker-webpack-plugin' {
   const plugin: any;
-  export = plugin
+  export = plugin;
+}
+declare module 'eslint-plugin-no-use-extend-native' {
+  import type {Eslint} from 'eslint';
+  const plugin: Eslint.Plugin;
+  export = plugin;
+}
+declare module 'eslint-plugin-array-func' {
+  import type {Eslint} from 'eslint';
+  const plugin: Eslint.Plugin;
+  export = plugin;
+}
+declare module 'eslint-plugin-github' {
+  import type {Eslint} from 'eslint';
+  const plugin: Eslint.Plugin;
+  export = plugin;
+}
+declare module 'eslint-plugin-no-jquery' {
+  import type {Eslint} from 'eslint';
+  const plugin: Eslint.Plugin;
+  export = plugin;
+}
+declare module '@eslint-community/eslint-plugin-eslint-comments' {
+  import type {Eslint} from 'eslint';
+  const plugin: Eslint.Plugin;
+  export = plugin;
+}
+declare module '@eslint/compat' {
+  const value: any;
+  export const fixupPluginRules = value;
 }
