+++ web_src/js/types.ts
@@ -4,7 +4,7 @@ export type MentionValue = {
   name: string,
   fullname: string,
   avatar: string,
-}
+};
 
 export type Config = {
   appUrl: string,
@@ -20,7 +20,7 @@ export type Config = {
   mentionValues?: MentionValue[],
   mermaidMaxSourceCharacters: number,
   i18n: Record<string, string>,
-}
+};
 
 export type IntervalId = ReturnType<typeof setInterval>;
 
@@ -35,21 +35,21 @@ export type RequestOpts = {
 export type RepoOwnerPathInfo = {
   ownerName: string,
   repoName: string,
-}
+};
 
 export type IssuePathInfo = {
   ownerName: string,
   repoName: string,
   pathType: string,
   indexString?: string,
-}
+};
 
 export type IssuePageInfo = {
   repoLink: string,
   repoId: number,
   issueNumber: number,
   issueDependencySearchType: string,
-}
+};
 
 export type Issue = {
   id: number;
@@ -65,6 +65,6 @@ export type Issue = {
 export type FomanticInitFunction = {
   settings?: Record<string, any>,
   (...args: any[]): any,
-}
+};
 
 export type GitRefType = 'branch' | 'tag';
