+++ web_src/js/utils/glob.test.ts
@@ -117,7 +117,7 @@ test('GlobCompiler', async () => {
   for (const c of golangCases) {
     const compiled = globCompile(c.pattern, c.separators);
     const msg = `pattern: ${c.pattern}, input: ${c.input}, separators: ${c.separators || '(none)'}, compiled: ${compiled.regexpPattern}`;
-    // eslint-disable-next-line @vitest/valid-expect -- Unlike Jest, Vitest supports a message as the second argument
+    // eslint-disable-next-line vitest/valid-expect -- Unlike Jest, Vitest supports a message as the second argument
     expect(compiled.regexp.test(c.input), msg).toBe(c.matched);
   }
